client-angular2/client/src/app/+account/+dashboard/shared/objectives/objectives.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { ObjectivesComponent } from './objectives.component';

describe('Component: Objectives', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [ObjectivesComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([ObjectivesComponent],
      (component: ObjectivesComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(ObjectivesComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(ObjectivesComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-objectives></app-objectives>
  `,
  directives: [ObjectivesComponent]
})
class ObjectivesComponentTestController {
}

